requirements.txt
# ===================================
# WEB FRAMEWORK
# ===================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# ===================================
# DATABASE
# ===================================
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
asyncpg==0.29.0

# ===================================
# DATA SCIENCE & ML
# ===================================
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
scipy==1.11.4

# ===================================
# DEEP LEARNING (Optional - PyTorch)
# ===================================
torch==2.1.1
torchvision==0.16.1

# ===================================
# MODEL PERSISTENCE
# ===================================
joblib==1.3.2

# ===================================
# UTILITIES
# ===================================
python-dotenv==1.0.0
requests==2.31.0
httpx==0.25.2

# ===================================
# LOGGING & MONITORING
# ===================================
loguru==0.7.2

# ===================================
# TESTING
# ===================================
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# ===================================
# DATA MIGRATION & SYNC
# ===================================
alembic==1.13.0
schedule==1.2.0
